requirements.txt
itchat
jinja2
pyyaml
gitpython